@@ -1,6 +1,6 @@
 package com.openclassrooms.mddapi.models;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table(name = "subscriptions", uniqueConstraints = {
